Move tools into PyCanZE and improve scan_car
