+++ PyCanZE/AGENTS.md
@@ -1,11 +1,18 @@
 # PyCanZE
 
-PyCanZE is an experimental Python port of the CanZE project. It aims to parse the CSV asset files used by the Android app and provide a flexible foundation for polling vehicle data via standard ELM327-compatible interfaces. The long-term goal is to offer command-line and library tools that can publish decoded values to automation systems such as MQTT/Home Assistant.
+PyCanZE is a Python spin-off of the CanZE project, an Android/iOS graphical
+tool for Renault vehicles. The CSV data files from the original app are copied
+here to support Python command-line tooling and experimentation.
+
+The immediate aim is to poll selected vehicle registers and publish decoded
+values over MQTT. A GUI similar to the original app may be built later.
 
 ## Current state
 
-- Basic models and a CSV asset parser are implemented in `parser.py` and `models.py`.
-- Raw UDS communication helpers exist in `uds.py` but there is no generic polling client yet.
-- No MQTT interface, command-line utility or test suite has been added so far.
+- Dataclasses and CSV parsers live in `models.py` and `parser.py`.
+- A minimal UDS client in `uds.py` powers example utilities in `tools/` and
+  `Testing/`.
+- No MQTT interface or test suite is available yet.
 
 Contributions are welcome as this module is under active development.
+
