+++ PyCanZE/Testing/canze_dongle_selftest.py
