+++ PyCanZE/Testing/logs/zoe_2025-08-27_22-26-46.log
